// src/generated/api/models/api-v1-auth-login-post200-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SmartSeller Backend API
 * API documentation for the SmartSeller Backend service - E-commerce Management Platform.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ApiV1AuthLoginPost200ResponseData } from './api-v1-auth-login-post200-response-data';
// May contain unused imports in some cases
// @ts-ignore
import type { ApiV1AuthLoginPost200ResponseMeta } from './api-v1-auth-login-post200-response-meta';

/**
 * 
 * @export
 * @interface ApiV1AuthLoginPost200Response
 */
export interface ApiV1AuthLoginPost200Response {
    /**
     * 
     * @type {boolean}
     * @memberof ApiV1AuthLoginPost200Response
     */
    'success'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ApiV1AuthLoginPost200Response
     */
    'message'?: string;
    /**
     * 
     * @type {ApiV1AuthLoginPost200ResponseData}
     * @memberof ApiV1AuthLoginPost200Response
     */
    'data'?: ApiV1AuthLoginPost200ResponseData;
    /**
     * 
     * @type {ApiV1AuthLoginPost200ResponseMeta}
     * @memberof ApiV1AuthLoginPost200Response
     */
    'meta'?: ApiV1AuthLoginPost200ResponseMeta;
}