// src/generated/api/models/api-v1-auth-google-login-get200-response-data.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SmartSeller Backend API
 * API documentation for the SmartSeller Backend service - E-commerce Management Platform.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ApiV1AuthGoogleLoginGet200ResponseData
 */
export interface ApiV1AuthGoogleLoginGet200ResponseData {
    /**
     * The Google OAuth URL to redirect to
     * @type {string}
     * @memberof ApiV1AuthGoogleLoginGet200ResponseData
     */
    'redirect_url'?: string;
    /**
     * CSRF protection state parameter
     * @type {string}
     * @memberof ApiV1AuthGoogleLoginGet200ResponseData
     */
    'state'?: string;
}